Extract custom hooks from view components
Move repeated network/stream/state patterns into three reusable hooks
so the view components become smaller and easier to split later.

- use-authorized-fetch: wraps retryOnceAfterUnauthorized + router.refresh
- use-sync-stream: wraps subscribeToSyncStream lifecycle in useEffect
- use-activity-detail: manages detail-overlay fetch, abort, toggle, prune

Closes #371
